move `org.discu2.forum` to `org.discu2.forum.core`
