+++ core/src/main/java/org/discu2/forum/core/Core.java
@@ -1,0 +1,16 @@
+package org.discu2.forum.core;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+
+@ComponentScan("org.discu2.forum")
+@SpringBootApplication(exclude= {UserDetailsServiceAutoConfiguration.class})
+public class Core {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Core.class, args);
+    }
+
+}
