+++ core/src/main/resources/application-dev.properties
