+++ core/src/main/resources/application-prod.properties
