+++ core/src/test/java/org/discu2/forum/CoreTests.java
@@ -1,0 +1,19 @@
+package org.discu2.forum;
+
+import org.junit.jupiter.api.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@ActiveProfiles("test")
+@SpringBootTest
+class CoreTests {
+
+    @Test
+    void contextLoads() {
+        System.out.println("Testing...🥺");
+    }
+
+}
