--- forum/src/main/java/org/discu2/forum/Forum.java
@@ -1,14 +1,0 @@
-package org.discu2.forum;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration;
-
-@SpringBootApplication(exclude= {UserDetailsServiceAutoConfiguration.class})
-public class Forum {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Forum.class, args);
-    }
-
-}
