--- forum/src/main/java/org/discu2/forum/config/MethodSecurityConfig.java
@@ -1,26 +1,0 @@
-package org.discu2.forum.config;
-
-import lombok.AllArgsConstructor;
-import org.discu2.forum.service.ForumPermissionService;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.access.expression.method.DefaultMethodSecurityExpressionHandler;
-import org.springframework.security.access.expression.method.MethodSecurityExpressionHandler;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.method.configuration.GlobalMethodSecurityConfiguration;
-
-@Configuration
-@AllArgsConstructor
-@EnableGlobalMethodSecurity(prePostEnabled = true, securedEnabled = true, jsr250Enabled = true)
-public class MethodSecurityConfig extends GlobalMethodSecurityConfiguration {
-
-    private ForumPermissionService forumPermissionService;
-
-    @Override
-    protected MethodSecurityExpressionHandler createExpressionHandler() {
-
-        var expressionHandler = new DefaultMethodSecurityExpressionHandler();
-        expressionHandler.setPermissionEvaluator(forumPermissionService);
-
-        return expressionHandler;
-    }
-}
