--- forum/src/test/java/org/discu2/forum/ForumTests.java
@@ -1,19 +1,0 @@
-package org.discu2.forum;
-
-import org.junit.jupiter.api.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-@RunWith(SpringJUnit4ClassRunner.class)
-@ActiveProfiles("test")
-@SpringBootTest
-class ForumTests {
-
-    @Test
-    void contextLoads() {
-        System.out.println("Testing...🥺");
-    }
-
-}
